// examples/convex/runtimes/convex/_generated/api.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import { makeActionReference, makeMutationReference, makeQueryReference } from "convex/browser";
import type { Doc, Id, JsonValue } from "./dataModel.js";

export const api = {
  digests: {
    list: makeQueryReference<{}, Doc<"digests">[]>("digests:list", "public"),
    hashWithDefaultRuntime: makeActionReference<{
  "text": string;
}, Id<"digests">>("digests:hashWithDefaultRuntime", "public")
  },
  nodeDigests: {
    hashWithNodeRuntime: makeActionReference<{
  "text": string;
}, Id<"digests">>("nodeDigests:hashWithNodeRuntime", "public")
  },
  shareIds: {
    create: makeMutationReference<{}, Id<"shareIds">>("shareIds:create", "public"),
    list: makeQueryReference<{}, Doc<"shareIds">[]>("shareIds:list", "public")
  }
};

export const internal = {
  digests: {
    store: makeMutationReference<{
  "runtime": "default" | "node";
  "algorithm": string;
  "input": string;
  "output": string;
}, Id<"digests">>("digests:store", "internal")
  }
};