// examples/convex/runtimes/convex/_generated/dataModel.d.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import type { GenericId } from "convex/values";

export type JsonValue = null | boolean | number | string | JsonValue[] | { [key: string]: JsonValue };

export type TableNames = "digests" | "shareIds";

export type Id<TableName extends string> = GenericId<TableName>;

type DocumentByTable = {
  "digests": {
    _id: Id<"digests">;
    _creationTime: number;
    _updateTime: number;
    "runtime": "default" | "node";
    "algorithm": string;
    "input": string;
    "output": string;
    "createdAt": number;
  };
  "shareIds": {
    _id: Id<"shareIds">;
    _creationTime: number;
    _updateTime: number;
    "id": string;
    "createdAt": number;
  };
};

type IndexNamesByTable = {
  "digests": "by_created_at";
  "shareIds": "by_created_at";
};

export type Doc<TableName extends TableNames> = DocumentByTable[TableName];

export type DataModel = {
  [TableName in TableNames]: {
    document: DocumentByTable[TableName];
    fieldPaths: keyof DocumentByTable[TableName];
    indexes: IndexNamesByTable[TableName];
  };
};